pyproject.toml
[tool.black]
line-length = 120
target-version = ["py38"]


[tool.poetry]
name = "scargo"
version = "0.0.1"
description = "Code for scargo project."
authors = ["ProteinQure team <team@proteinqure.com>"]
license = "Proprietary"


[[tool.poetry.source]]
name = "ProteinQure"
url = "https://pypi.proteinqure.com/simple"


[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
pyyaml = "^5.4.1"
pytest = "^6.2.2"


[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pytest-console-scripts = "^1.1.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
scargo = 'scargo.cli:run'
